#23 refactor Manager.jsx
* #23 create a reusable useWebSocket component

* #23 remove unused variable ws in Manager.jsx

* #23 extract Snackbar error from Manager.jsx

* #23 extract methods from Manager component and put them under Manager/ folder

* #23 edit WebSocketHandler with synchronized and to be able to handle null
